Do away with the concept of cert & key defines
We're going to do everything via the cert define as it makes logical
sense. Otherwise we were going to be replicating basically all of the
cert logic into the key define. A couple of extra flags to cert and it
can be used for deploying the keys as well (which is the default).

Some minor updates to the README.md but it still needs more information
and work but that will really come from real-life tests at this point.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>